Create MainDisplayItem class
