+++ QviKD/MainWindow.xaml
@@ -4,7 +4,7 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d"
-        Title="QviKD" Height="480" Width="800" MinWidth="448" MinHeight="400" Initialized="Window_Initialized" Loaded="Window_Loaded">
+        Title="QviKD" Height="480" Width="800" MinWidth="448" MinHeight="400" Loaded="Window_Loaded">
     <DockPanel LastChildFill="True">
         <Grid DockPanel.Dock="Top" Background="#FF202020" Height="48"></Grid>
         <StatusBar DockPanel.Dock="Bottom" Height="24" Background="#FF202020"></StatusBar>
